pyproject.toml
[project]
name = "medic_analysis"
description = "Various files and analysis for MEDIC paper"
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT License" }
authors = [{ name = "Andrew Van", email = "vanandrew@wustl.edu" }]
keywords = ["neuroimaging"]
classifiers = ["License :: OSI Approved :: MIT License", "Programming Language :: Python :: 3"]
urls = { github = "https://github.com/vanandrew/medic_analysis" }
version = "0.0.1"
dynamic = ["entry-points"]
dependencies = [
  "nibabel >= 4.0.2",
  "indexed-gzip >= 1.7.0",
  "pybids >= 0.15.1",
  "memori >= 0.2.1",
  "warpkit >= 0.0.1",
  "numpy >= 1.23.3",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
zip-safe = true

[tool.setuptools.packages.find]
exclude = ["tests"]

[tool.black]
line-length = 120
target-version = ["py37", "py38", "py39", "py310"]
